refactor: Extract <DetourDropdownItem />
